Proceesing and Running model
The data_processing.py file is for the gaussian filter. The others have
resize appended to the name because these ones are the ones where the
images were resized. The run_models.py is the one used for the mixture
of alphabets. The other two have lat for Latin and cyr for Cyrillic
appended to the names.